Fixes after second review
