+++ Boa.Constrictor.UnitTests/Screenplay/Conditions/Enumerations/EveryItemTest.cs
