+++ Boa.Constrictor.UnitTests/Screenplay/Conditions/Enumerations/FirstItemTest.cs
