+++ Boa.Constrictor.UnitTests/Screenplay/Conditions/Enumerations/LastItemTest.cs
