+++ Boa.Constrictor/Screenplay/Conditions/Enumerations/AtLeastOneItem.cs
@@ -48,8 +48,8 @@ namespace Boa.Constrictor.Screenplay
         ///     ToString override.
         /// </summary>
         /// <returns></returns>
-        public override string ToString() => $"is an enumerable of type {typeof(T)} and at least one item {Condition}";
-
+        public override string ToString() => $"is an IEnumerable<{typeof(T)}> where at least one item {Condition}";
+        
         #endregion
     }
 }
