+++ Boa.Constrictor/Screenplay/Conditions/Enumerations/EveryItem.cs
@@ -48,7 +48,7 @@ namespace Boa.Constrictor.Screenplay
         ///     ToString override.
         /// </summary>
         /// <returns></returns>
-        public override string ToString() => $"is an enumerable of type {typeof(T)} and every item {Condition}";
+        public override string ToString() => $"is an IEnumerable<{typeof(T)}> where every item {Condition}";
 
         #endregion
     }
