+++ Boa.Constrictor/Screenplay/Conditions/Enumerations/FirstItem.cs
@@ -47,7 +47,7 @@ namespace Boa.Constrictor.Screenplay
         ///     ToString override.
         /// </summary>
         /// <returns></returns>
-        public override string ToString() => $"is an enumerable of type {typeof(T)} and the first item {Condition}";
+        public override string ToString() => $"is an IEnumerable<{typeof(T)}> where the first item {Condition}";
 
         #endregion
     }
