+++ Boa.Constrictor/Screenplay/Conditions/Enumerations/LastItem.cs
@@ -49,7 +49,7 @@ namespace Boa.Constrictor.Screenplay
         ///     ToString override.
         /// </summary>
         /// <returns></returns>
-        public override string ToString() => $"is an enumerable of type {typeof(T)} and the last item {Condition}";
+        public override string ToString() => $"is an IEnumerable<{typeof(T)}> where the last item {Condition}";
 
         #endregion
     }
